Add benchmark for AbsUrlInXml
And a general test cleanup in /transform.

See #894
